requirements.txt
appdirs==1.4.3
Django==1.7
djangorestframework==3.1.1
packaging==16.8
pyparsing==2.2.0
six==1.10.0
